streamlined the flow
