+++ pixi.toml
@@ -54,6 +54,7 @@ langchain-deepseek = "*"
 GIT_TOKEN_CHECK = { cmd = "os.getenv(GITHUB_TOKEN)" }
 DJ_package = { cmd = "pip install --editable . --use-pep517" }
 DJ_pre_commit = { cmd = "pre-commit install" }
+DJ_list = { cmd = "python cli.py"}
 # applications
 DJ_dagster = { cmd = "dagit -f process.py", depends-on = "DJ_package", cwd = "analytics_framework/pipeline" }
 DJ_panel_app = { cmd = "python stock_price_twilio_app.py", depends-on = "DJ_package", cwd = "analytics_framework/dashboard" }
